Adjust input array for Big O test case
Increase number of elements in input array from 1000 to 10000.

Previously benchmarks showed that an input array of 1000 elements was
too small to make a simple nested for loop structure (that is to say, a
Big O(n ^2) time complexity) time out. The increase to 10000 not only
makes the test fail as intended, but gives padding to account for more
powerful machines running the tests inherently faster.

The previous code also shows a comment that seems to show the tests
were intended to equal 10000, which makes it seem as if 10000 was the
intended value all along.