Adjust test where all items are negative
Change the expected solution of the test when all items are negative to
equal the maximum element. In the case of the provided array, the answer
would be -1. This aligns with how the question is phrased, and is often
represented on algorithm sites.

Also change the input array so that the maximum element is a middle
element. This accounts for solutions that set the initial maximum to the
first element of the input array.

An alternate solution would be to change the wording of the problem to
only allow for sums above or equal to zero.